% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome.R
\name{chromosome}
\alias{chromosome}
\title{chromosome}
\usage{
chromosome(bamfile, option)
}
\arguments{
\item{bamfile}{path to bam file}

\item{option}{either count, rate or length enter as string}
}
\description{
chromosome
}
