% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{cover_age}
\alias{cover_age}
\title{coverage rate: mean, median and sd}
\usage{
cover_age(bamfile)
}
\arguments{
\item{bamfile}{the path to the bam file}
}
\value{
coverage rate mean, median and sd matrix
}
\description{
coverage rate: mean, median and sd
}
