% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_map.R
\name{mult_map}
\alias{mult_map}
\title{multiple mapping}
\usage{
mult_map(bamfile, quality = 30, option)
}
\arguments{
\item{bamfile}{the path to bam file}

\item{quality}{set default of mapping quality standard to 30}

\item{option}{either count or rate}
}
\description{
multiple mapping
}
