% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_reads.R
\name{unmapped_reads}
\alias{unmapped_reads}
\title{return name of unmapped reads}
\usage{
unmapped_reads(bamfile)
}
\arguments{
\item{bamfile}{the path to the bam file}
}
\description{
return name of unmapped reads
}
