% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmapped_reads.R
\name{unmapped_reads}
\alias{unmapped_reads}
\title{return name of unmapped reads}
\usage{
unmapped_reads(bamfile, option)
}
\arguments{
\item{bamfile}{the path to the bam file}

\item{option}{'count' or 'rate' gives either count of unmapped reads or the rate of unmapped reads}
}
\description{
return name of unmapped reads
}
